Add jquery typings to project
Remove protractor typings
As detailed in this issue,
https://github.com/DefinitelyTyped/DefinitelyTyped/issues/2734,
the protractor typings are incompatible with jQuery. The best
workaround suggested is to move integration tests via protractor
and selenium to a different project. Since we will likely not
worry about integration tests in this project, I just removed
the protractor typings.