Specify Accept header for JSON response
Because of a bug in Firefox, XHR requests that do not specify an Accept
header get one requesting for an HTML response by default. The Django
Rest Framework uses the Accept header to specify how to serialize the
response, which means that we previously had to manually specify
"?format=json" in each request. Using the Accept header for this better
fits the HTTP style, frees up space in the request URL, and bypasses the
Firefox bug entirely.