Read time aggregation from JSON rather than title
Previously we detected the indicator's aggregation level (Yearly,
monthly, day) from the title of the indicator, rather than the
time_aggregation object within the JSON response. This causes indicators
that don't feature that word to be interpreted incorrectly, as well as
posing translation concerns for the future. By using the JSON response
directly we alleviate those problems.