Add transition delay on tooltip
Since we are using these tooltips more as helper text on buttons, adding a slight delay is a nice touch. When there is no delay, the tooltips can feel overwhelming. If I know what a button does and move my mouse over it and click it immediately, I don't need a tooltip in that situation. These tooltips are more useful for one time reference while a user learns how to use the application.