Remove unused unit selector
Remove unit selector in add/edit form that will never appear.
It is in a section that only shows when adding a new project, but itself will only appear if there is a chart selected.
There will never be a chart selected already on new project creation.

Makes room for dataset selector instead.