Mostly remove use of lodash
In favor of ES6 builtins

Only one that couldn't be effectively removed
was _.cloneDeep(). Attempted naive replacements
via:
1. const clone = Object.assign({}, original);
2. const clone = JSON.parse(JSON.stringify(original));

1 was attempted because we really only need to trigger change
detection. But it didn't seem to do so.

2 was attempted as a suggested alternative in the JS docs,
but as expected it reverts objects to their 'toString'
representations, because we have to stringify. So it doesn't
properly preserve the date objects in the objects we're cloning.

Ref: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign#Deep_Clone

Since I didn't have any luck removing, I swapped the
node dep to specifically target lodash.clonedeep. This
prevents desire to use lodash for features provided
by ES6 builtins since we no longer include the whole
library.