Update styling to accommodate wider scenario toggle buttons
In the newest version of the `climate-change-components` the scenario
toggle now uses "Low emissions" and "High emissions" instead of the more
cryptic "RCP 4.5" and "RCP 8.5" for labels. This is more helpful, but the
longer text messed up the styling.

Increasing the width on the new project form, and allowing the navbar
header to wrap to multiple lines fixes the styling while retaining the
more helpful labeling.