Allow creating projects for locations that don't have an ingested City
Rather than using the city endpoint for the autocomplete in the new project
form and the lab header, we now use a Google geocoder, and make a separate AJAX
call to the API to determine if that lat/lon has data.

This will allow us to support a broader number of locations than we do
currently.

Closes #333
Closes #266
Closes #251