Fix race condition in displaying models modal button
Changing the timing of endpoints on the backend confirmed my hypothesis
that this issue was due to a data race. If the list of models loads
*after* the list of datasets, the models button loads fine, however when
the order is reversed the button doesn't fully load.

This is because the models button needs to know the current dataset to
know which models are available, but we first load available datasets and
then set a default dataset, leaving a period of time when the dataset is
unset. This only an issue on the new project page, as on the lab page there
is a preconfigured dataset, which is why we can avoid a guard there.

Closes #305